requirements.txt
# ---- Core Web Framework ----
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
starlette==0.36.3

# ---- Database ----
pymongo==4.6.1

# ---- Auth & Security ----
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1

# ---- Resume Parsing ----
pdfplumber==0.10.3
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10

# ---- NLP / AI (optional but safe) ----
spacy==3.7.4
openai==1.54.4
httpx==0.27.2

# ---- Utilities ----
requests==2.31.0
python-dateutil==2.9.0.post0